backend-python/requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=1.0.4
langgraph-checkpoint>=2.0.0
openai>=1.6.1
httpx>=0.26.0
python-multipart>=0.0.6